+++ src/test/java/net/boyechko/pdf/autoa11y/fixes/WrapParagraphRunInListTest.java
@@ -60,6 +60,64 @@ class WrapParagraphRunInListTest extends PdfTestBase {
     }
 
     @Test
+    void nestsSublistIntoPrecedingListItem() throws Exception {
+        try (PdfDocument pdfDoc = new PdfDocument(new PdfWriter(testOutputStream()))) {
+            pdfDoc.setTagged();
+            pdfDoc.addNewPage();
+            PdfStructTreeRoot root = pdfDoc.getStructTreeRoot();
+            PdfStructElem document = new PdfStructElem(pdfDoc, PdfName.Document);
+            root.addKid(document);
+
+            PdfStructElem list = new PdfStructElem(pdfDoc, PdfName.L);
+            document.addKid(list);
+            PdfStructElem li = new PdfStructElem(pdfDoc, PdfName.LI);
+            list.addKid(li);
+            PdfStructElem lBody = new PdfStructElem(pdfDoc, PdfName.LBody);
+            li.addKid(lBody);
+            PdfStructElem itemP = new PdfStructElem(pdfDoc, new PdfName("P"));
+            lBody.addKid(itemP);
+
+            PdfStructElem sub1 = new PdfStructElem(pdfDoc, new PdfName("P"));
+            PdfStructElem sub2 = new PdfStructElem(pdfDoc, new PdfName("P"));
+            document.addKid(sub1);
+            document.addKid(sub2);
+
+            DocContext ctx = new DocContext(pdfDoc);
+            WrapParagraphRunInList fix =
+                    new WrapParagraphRunInList(document, List.of(sub1, sub2), list);
+            fix.apply(ctx);
+
+            assertEquals(
+                    "Document[L[LI[LBody[P[], L[LI[LBody[P[]]], LI[LBody[P[]]]]]]]]",
+                    StructTree.toRoleTree(document).toString());
+        }
+    }
+
+    @Test
+    void fallsBackToSiblingListWhenNestTargetHasNoLBody() throws Exception {
+        try (PdfDocument pdfDoc = new PdfDocument(new PdfWriter(testOutputStream()))) {
+            pdfDoc.setTagged();
+            pdfDoc.addNewPage();
+            PdfStructTreeRoot root = pdfDoc.getStructTreeRoot();
+            PdfStructElem document = new PdfStructElem(pdfDoc, PdfName.Document);
+            root.addKid(document);
+
+            PdfStructElem list = new PdfStructElem(pdfDoc, PdfName.L);
+            document.addKid(list);
+
+            PdfStructElem sub1 = new PdfStructElem(pdfDoc, new PdfName("P"));
+            document.addKid(sub1);
+
+            DocContext ctx = new DocContext(pdfDoc);
+            WrapParagraphRunInList fix = new WrapParagraphRunInList(document, List.of(sub1), list);
+            fix.apply(ctx);
+
+            assertEquals(
+                    "Document[L[], L[LI[LBody[P[]]]]]", StructTree.toRoleTree(document).toString());
+        }
+    }
+
+    @Test
     void doesNothingWithRegularParagraphs() throws Exception {
         try (PdfDocument pdfDoc = new PdfDocument(new PdfWriter(testOutputStream()))) {
             pdfDoc.setTagged();